pyproject.toml
[project]
name = "mcp-raganything"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.31.0",
    "docling>=2.64.0",
    "fastapi>=0.124.0",
    "authlib>=1.6.9",
    "fastmcp>=2.14.3",
    "httpx>=0.27.0",
    "lightrag-hku>=1.4.9.8",
    "lightrag-hku[api]>=1.4.9.8",
    "mcp>=1.24.0",
    "minio>=7.2.18",
    "openai>=2.9.0",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "raganything>=1.2.8",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=xml:coverage.xml"
pythonpath = ["src"]


[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = ["E501", "B008"]  # E501: line too long (handled by formatter), B008: FastAPI Depends/File in defaults

[tool.ruff.lint.isort]
known-first-party = ["src"]
